+++ database/migrations/2019_07_12_164256_create_configures_table.php
@@ -18,9 +18,9 @@ class CreateConfiguresTable extends Migration
             $table->integer('distributor1')->unsigned()->commit('一级分销比例,百分比');
             $table->integer('distributor2')->unsigned()->commit('二级级分销比例,百分比');
             $table->integer('distributor3')->unsigned()->commit('三级分销比例,百分比');
-            $table->decimal('vip2_price_n', 5, 2)->unsigned()->commit('银牌会员无上级购买价格');
-            $table->decimal('vip2_price_y', 5, 2)->unsigned()->commit('银牌会员有上级购买价格');
-            $table->integer('vip3_price')->unsigned()->commit('代理价格,线下缴费');
+            $table->decimal('vip2_price_n', 9, 2)->unsigned()->commit('银牌会员无上级购买价格');
+            $table->decimal('vip2_price_y', 9, 2)->unsigned()->commit('银牌会员有上级购买价格');
+            $table->decimal('vip3_price', 9, 2)->unsigned()->commit('代理价格,线下缴费');
             $table->integer('invite_copper')->unsigned()->commit('邀请用户得铜币数');
             $table->integer('zan_copper')->unsigned()->commit('点赞得铜币数');
             $table->integer('buy_vip2_self')->unsigned()->commit('购买银牌会员自己得银币数');
